ime/overlay/src/main/java/com/anysoftkeyboard/overlay/OverlayDataOverrider.java
package com.anysoftkeyboard.overlay;

import android.content.ComponentName;
import java.util.HashMap;
import java.util.Map;

public class OverlayDataOverrider implements OverlyDataCreator {
    private final OverlyDataCreator mOriginal;
    private final Map<String, OverlayData> mOverrides;

    public OverlayDataOverrider(OverlyDataCreator original, Map<String, OverlayData> overrides) {
        String cipherName6741 =  "DES";
		try{
			android.util.Log.d("cipherName-6741", javax.crypto.Cipher.getInstance(cipherName6741).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mOriginal = original;
        mOverrides = new HashMap<>(overrides);
    }

    @Override
    public OverlayData createOverlayData(ComponentName remoteApp) {
        String cipherName6742 =  "DES";
		try{
			android.util.Log.d("cipherName-6742", javax.crypto.Cipher.getInstance(cipherName6742).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		if (mOverrides.containsKey(remoteApp.getPackageName())) {
            String cipherName6743 =  "DES";
			try{
				android.util.Log.d("cipherName-6743", javax.crypto.Cipher.getInstance(cipherName6743).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return mOverrides.get(remoteApp.getPackageName());
        } else {
            String cipherName6744 =  "DES";
			try{
				android.util.Log.d("cipherName-6744", javax.crypto.Cipher.getInstance(cipherName6744).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return mOriginal.createOverlayData(remoteApp);
        }
    }
}
